// AdministracionApp/src/app/models/user.model.ts
export class User {
  email: string;
  password: string;
  nombre: string;
  apellido: string;
  dni: string;
  uid?: string;
  tipoUsuario: string;
  foto: string;
  sexo: string;
  codigo_100?: number;
  codigo_50?: number;
  codigo_10?: number;
  cargas_100?: number;
  cargas_500?: number;
  cargas_10?: number;
  credito?: number;
}